+++ common/src/main/java/club/iananderson/seasonhud/client/overlays/SeasonHUDOverlayCommon.java
@@ -2,7 +2,6 @@ package club.iananderson.seasonhud.client.overlays;
 
 import club.iananderson.seasonhud.Common;
 import club.iananderson.seasonhud.config.Config;
-import club.iananderson.seasonhud.config.Location;
 import club.iananderson.seasonhud.impl.minimaps.CurrentMinimap;
 import club.iananderson.seasonhud.impl.seasons.Calendar;
 import club.iananderson.seasonhud.impl.seasons.CurrentSeason;
@@ -12,7 +11,6 @@ import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.network.chat.MutableComponent;
 
 public class SeasonHUDOverlayCommon {
-
   public static void render(GuiGraphics seasonStack) {
     Minecraft mc = Minecraft.getInstance();
     MutableComponent seasonCombined = CurrentSeason.getInstance(mc).getSeasonHudText();
@@ -26,11 +24,9 @@ public class SeasonHUDOverlayCommon {
     int yOffset = (int) ((Config.hudY.get() + offsetDim) / guiSize);
     int stringWidth = mc.font.width(seasonCombined);
 
-    if (Config.enableMod.get() && (CurrentMinimap.noMinimapLoaded() || (Services.MINIMAP.currentMinimapHidden()
-        && Config.showDefaultWhenMinimapHidden.get()) || !Config.enableMinimapIntegration.get())) {
-      Location hudLoc = Config.hudLocation.get();
-
-      switch (hudLoc) {
+    if (Config.enableMod.get() && (CurrentMinimap.noMinimapLoaded() || (Services.MINIMAP.allMinimapsHidden()
+        && Config.showDefaultWhenMinimapHidden.get()))) {
+      switch (Config.hudLocation.get()) {
         case TOP_LEFT -> {
           x = offsetDim;
           y = offsetDim;